Add a delay in onUnbind
After a reconfiguration, the service has no time to destroy himself that it already started. That cause an error because mService reference a dead thread.

This solution seems to work because onCreate of Service is call when it should but I can't see log of stopService and onDestroy. It's really weird but it's seems to work